split out routines for splitting graphs
