+++ src/weather_model_graphs/__init__.py
@@ -1,5 +1,4 @@
-from . import create, save, visualise
-from .networkx_utils import (
-    replace_node_labels_with_unique_ids,
-    split_graph_by_edge_attribute,
-)
+from . import create, visualise
+from .labelling import replace_node_labels_with_unique_ids
+from .save import pyg
+from .split import split_graph_by_edge_attribute
