+++ src/weather_model_graphs/create/grid/grid.py
@@ -1,6 +1,6 @@
 import networkx
 
-from ...networkx_utils import prepend_node_index
+from ...labelling import prepend_node_index
 
 
 def create_grid_graph_nodes(xy, level_id=-1):
